Backported bugfixes from 1.0 branch.
git-svn-id: https://opentk.svn.sourceforge.net/svnroot/opentk/trunk@2588 ebc5dd9b-fb1d-0410-b6f8-d24c324e9604

+++ Source/OpenTK/Graphics/OpenGL/GLHelper.cs
@@ -729,7 +729,7 @@ namespace OpenTK.Graphics.OpenGL
 
         public static void TexCoordPointer(int size, TexCoordPointerType type, int stride, int offset)
         {
-            TexCoordPointer(size, type, stride, offset);
+            TexCoordPointer(size, type, stride, (IntPtr)offset);
         }
 
         public static void VertexAttribPointer(int index, int size, VertexAttribPointerType type, bool normalized, int stride, int offset)
